+++ src/v_draw.cpp
@@ -49,7 +49,7 @@
 
 // [RH] Stretch values to make a 320x200 image best fit the screen
 // without using fractional steppings
-float CleanXfac, CleanYfac;
+int CleanXfac, CleanYfac;
 
 // [RH] Effective screen sizes that the above scale values give you
 int CleanWidth, CleanHeight;
