+++ src/v_video.h
@@ -47,7 +47,8 @@
 #include "r_data.h"
 
 // [BC] Make CleanX/Yfac floats in Skulltag.
-extern float CleanXfac, CleanYfac;
+// [BB] Changed CleanX/Yfac back to int in Skulltag. I see no reason why they should be float. This also removes some warnings.
+extern int CleanXfac, CleanYfac;
 extern int CleanWidth, CleanHeight;
 extern int DisplayWidth, DisplayHeight, DisplayBits;
